Add files via upload
This Excel project aims to answer the following questions:

How do sales for the various product lines trend over time?
What are the most popular product offerings by sales?
How does the company do in international markets?
Which companies are high contributors to success?

Brief Overview:

Taking this sales data I perform data cleaning processes and manipulation of the data to prepare the data for analysis. I use Pivot Tables and Data visualizations to draw insights to create a dashboard that answers each question and allows for users to gain meaningful insights. Provided is a cleaning document that will go in depth to the processes and thinking for each action.